Basic handling of tradera account consent
This is the initial commit. It sets up a dotnet solution like this:
- dostadning.domain: Defines models, procedures and validity of state
- dostadning.records: Implements data storage. Initially postgresql for
dostadning.domain.
- dostadning.soap: Implements communication via soap to external
services. Initially traderas soap-api for dostadning.domain
- dostadning.cli: A one-to-one mapping between procedures in
dostadning.domain for integration testing purposes.
- dostadning.unit: Unit tests

The commit implements handling of tradera consent.
1. We issue an id for a user of dostadning
2. The client can request that we associate this id with a tradera
alias. If we can confirm that this alias indeed is a tradera alias we
comply and issue a UUID for the user to confirm consent for dostadning
to use the restricted tradera api on behalf of that tradera alias.
3. At client explicit request of when it is needed we will fetch the
identity token through a call to the tradera api and record it.

For error handling we will aim for no exceptions thrown across projects.
Each implementeation there will catch it's own exceptions and return a
result object based on takes the Either type. This way avoid having many
references just for exception-types and rely on the type safety of
either instead of catching the right exeptions at the right time.

[investigate] It would be very nice to be able to load an call the procedures
interactively via fsharpi but the osx mono environment fails to load
entityframework core "File not found". I should investigate it further
later.