Use Observable instead of Task
We get error-propagation for free and we can write linq-qeries instead
of awaits for continuations, which is so much more easier to keep track
of. If we need to consume it with await we can just do that since there
is support for it. It we need to get at something synchronously we use
the Wait-function. This commit uses this in the cli-project. This commit
reduces LOC signiticantly with same functionality.