Create an Auction
We take some description of av acution that we verify has all the info
that tradera needs and we make a request to create an auction with
tradera. Back from tradera we get two keys: One to request updates on
the auction progress, the item id. The other is a handle on the request
itself. We can only verify that our command succeeded by polling tradea
with this id.

We need to have another go at scripting calls to tradea backend with f#.
The command line interface is just horrible. We have hardcoded a valid
tradera token in this commit